# DPL: introduce exponential backoff when idle (O2-626, O2-714)