--- a/Framework/Core/test/test_CallbackService.cxx
+++ b/Framework/Core/test/test_CallbackService.cxx
@@ -48,7 +48,7 @@ WorkflowSpec defineDataProcessing(ConfigContext const&)
          };
          ic.services().get<CallbackService>().set(CallbackService::Id::ClockTick, callback);
          return [count](ProcessingContext& ctx) {
-           if (*count > 1000) {
+           if (*count > 10) {
              ctx.services().get<ControlService>().readyToQuit(QuitRequest::All);
            }
          };