This should reduce the amount of CPU used by data processors when they are idle
while retaining the ability to increase back the rate, should it be needed.