@ktf 2020-02-24
@matthiasrichter this slightly changes the behaviour of the CallbackRegistry because the `Idle` callback is not accounted as "work" and therefore the exponential backoff reduces the amount of `Idle` callbacks to a 10Hz rate. I assume this is the behaviour we actually want as the `Idle` callbacks should be used for "quiet" periods.