@ktf 2020-02-25
Merging this for now, since it solves something people were (rightfully) ranting about. If we get any performance degradation we can simply drop the wait statement as a quick fix.