@Barthelemy 2020-02-27
I have tested it and the difference is staggering. I had a doubt if it was even running :)
Congrats and thank you