@knopers8 2020-02-27
I see no performance changes in Data Sampling larger than statistical errors.
Without:
```
Test date:              20-02-27 09:04:52
Latest commit:          bb35d24e8e4af35a8496f27d9cb91c9b4ff5ff69 Repeat in case of an error
Branch:                 data-sampling-benchmark-in-dev
Repetitions:            1
Test duration [s]:      300
Warm up cycles:         6
Free memory [B]:        62420000000
Memory soft limit [MB]: 12484
fraction       , payload size   , nb producers   , nb dispatchers , messages per second
           0.00,               1,               8,               1,            127369.1
           0.00,             256,               8,               1,            130894.5
           0.00,            1024,               8,               1,            129843.1
           0.00,            4096,               8,               1,            124769.8
           0.00,           16384,               8,               1,            125167.5
           0.00,           65536,               8,               1,             40541.9
           0.00,          262144,               8,               1,             21122.1
           0.00,         1048576,               8,               1,              4794.4
           0.00,         4194304,               8,               1,              1040.6
           0.00,        16777216,               8,               1,               241.4
           0.00,        67108864,               8,               1,                40.3
           0.00,       268435456,               8,               1,                10.2
           0.00,      1073741824,               8,               1,                 2.6
           1.00,               1,               8,               1,            130718.1
           1.00,             256,               8,               1,            122348.2
           1.00,            1024,               8,               1,            117129.3
           1.00,            4096,               8,               1,            115454.1
           1.00,           16384,               8,               1,             53729.9
           1.00,           65536,               8,               1,             19930.2
           1.00,          262144,               8,               1,              7797.7
           1.00,         1048576,               8,               1,              2368.7
           1.00,         4194304,               8,               1,               452.4
           1.00,        16777216,               8,               1,               117.1
           1.00,        67108864,               8,               1,                22.9
           1.00,       268435456,               8,               1,                 4.1
           1.00,      1073741824,               8,               1,                 1.0

```
With
```
Test date:              20-02-26 15:37:27
Latest commit:          bcd76d2afb0fde148544fa5ce6385424f984ead2 Merge branch 'data-sampling-benchmark-in-dev' of https://github.com/knopers8/AliceO2 into data-sampling-benchmark-in-dev
Branch:                 data-sampling-benchmark-in-dev
Repetitions:            1
Test duration [s]:      300
Warm up cycles:         6
Free memory [B]:        62495000000
Memory soft limit [MB]: 12499
fraction       , payload size   , nb producers   , nb dispatchers , messages per second
           0.00,               1,               8,               1,            130255.8
           0.00,             256,               8,               1,            128310.7
           0.00,            1024,               8,               1,            127742.2
           0.00,            4096,               8,               1,            128192.9
           0.00,           16384,               8,               1,            125784.5
           0.00,           65536,               8,               1,             45827.5
           0.00,          262144,               8,               1,             25603.7
           0.00,         1048576,               8,               1,              4883.2
           0.00,         4194304,               8,               1,               925.0
           0.00,        16777216,               8,               1,               262.1
           0.00,        67108864,               8,               1,                40.6
           0.00,       268435456,               8,               1,                10.3
           0.00,      1073741824,               8,               1,                 2.5
           1.00,               1,               8,               1,            119388.7
           1.00,             256,               8,               1,            127871.4
           1.00,            1024,               8,               1,            118204.6
           1.00,            4096,               8,               1,            111180.6
           1.00,           16384,               8,               1,             47352.9
           1.00,           65536,               8,               1,             23213.9
           1.00,          262144,               8,               1,              7207.0
           1.00,         1048576,               8,               1,              2261.6
           1.00,         4194304,               8,               1,               527.8
           1.00,        16777216,               8,               1,               110.6
           1.00,        67108864,               8,               1,                24.4
           1.00,       268435456,               8,               1,                 4.2
           1.00,      1073741824,               8,               1,                 1.2

```